Otap: better handle otap exchange with neighbors
When a node exchanges a scatchpad to one of its neighbors,
the dual mcu is not able to serve UART requests.
And dual mcu app for now doesn't discard old requests from
host. So better handle this situation to avoid detecting an
uart issue when it is just an OTAP exchange.

Also remove some redundant traces (especially printing
received buffers).