Automatically detect Microsoft Visual Studio NMAKE vs. GNU make
For the best user experience, typing "nmake" in the Microsoft Visual Studio
Developer Command Prompt should build the Wirepas C Mesh API library
successfully. The syntax of NMAKE and GNU make makefiles are sufficiently
different that a single makefile will not work for both.

This commit adds automatic detection of the make tool in use, and calls the
suitable tool-dependent makefile. Working NMAKE makefiles will be added in
another commit.