Add support for Darwin / macOS
This commit refactors the platform-specific code and splits it in POSIX,
Linux and Darwin (macOS) parts. Both Linux and Darwin use the POSIX
functionality, requiring just a small amount of OS-dependent code.

Build support with Clang etc. on macOS is also implemented. The build
platform is auto-detected by the GNU make makefile.