Add support for packed structs on Microsoft Visual Studio C compiler
The MSVC compiler does not understand the GCC __attribute__ syntax, but uses a
__pragma keyword instead. There's a pragma that can be used to enable struct
packing for a section of code instead of per type in the type definition,
like GCC does.

To support both GCC and MSVC in the same source code, an extra header file
called "compiler_ext.h" is introduced in this commit. It contains the
necessary macro trickery for managing packed structs.