Change Win32 build flags to be compatible with official Python release
In order to use the Wirepas C Mesh API library in a Python extension module,
it has to be built with flags that are compatible with the official Python
release for Windows. The default flag /MT ("link with LIBCMT.LIB") needs to
be changed to /MD ("link with MSVCRT.LIB").