Serial: try to reconnect to serial on write error
It is particularly usefull in case of USB connection as a
reboot of the board will close the serial link.
Reconnection is tried forever but there is a higher level
mechanism that close the program in case of too long interruption
in the communication with the sink (60s without successful poll).