Fix regression from previous commit
In case of long communication break with sink,
just exit the full process.
It is still not the best option as it should be
reported with a callback to the one that open the lib
but at least it fix the regression introduced with previous
version.
In fact if a sink was unplugged, after 60s, the polling thread was
exiting but nothing else, having the lib still open without any
polling.