Sink detection: better handle when sink do not respond
This is a workaround before a most robust solution.
Any succesfull poll request was considered as a heartbeat
form sink. Now consider any valid received message as a heartbeat.

In fact during otap exchanges, sink can stay silent for more than
30 seconds. And if two otap exchanges are following very quickly without
the oportunity of sending a poll confirm, sink can be silent for more
than 60s (lenght of the timeout).
This fix just reset the timeout with any messages to reduce the risk.
But the final solution should be on dual mcu side to send a busy signal
in such situation.