Data registration: register for all data
It is not that usefull to have a per EP data registration.
Dispatch will probably happen higher.
To simplify the code, make it as a compilation time flag.
By default if it is all data or nothing but it can still be change to
old behavior by defining REGISTER_DATA_PER_ENDPOINT flag.