Add support for scratchpad read primitive
This commit adds support for a new primitive in the Dual-MCU API that
allows reading out the stored scratchpad. This is needed to allow the
application MCU to use the scratchpad transfer mechanism for its own
firmware.

This commit also adds a new MSAP attribute for the stored scratchpad
size (attribute 14), so that the application process can query how big
the stored scratchpad is.

Test case for the scratchpad download has been added in test/test.c as
well.