Fix for deadlock in case timeout on communication is detected
The code detecting it is executed with the lock to access the sink held.
But as it calls the Platform_close() without releasing it, it generates
a deadlock while waiting for the polling thread to exit cleanly. In fact this
thread was never joined as it was waiting for the lock too..

So solution is to release the lock before calling Platform_Close.

Another solution was to not wait for the polling thread to finish as it
doesn't really have ressources attached, but still better to notify the thread
as in future it could have something to clean.